+++ src/main/java/com/habitpay/habitpay/domain/refreshtoken/application/RefreshTokenSearchService.java
@@ -1,0 +1,25 @@
+package com.habitpay.habitpay.domain.refreshtoken.application;
+
+import com.habitpay.habitpay.domain.member.application.MemberService;
+import com.habitpay.habitpay.domain.refreshtoken.dao.RefreshTokenRepository;
+import com.habitpay.habitpay.domain.refreshtoken.domain.RefreshToken;
+import com.habitpay.habitpay.global.config.jwt.TokenProvider;
+import com.habitpay.habitpay.global.config.jwt.TokenService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.NoSuchElementException;
+
+@RequiredArgsConstructor
+@Service
+@Slf4j
+public class RefreshTokenSearchService {
+
+    private final RefreshTokenRepository refreshTokenRepository;
+
+    public RefreshToken findByRefreshToken(String refreshToken) {
+        return refreshTokenRepository.findByRefreshToken(refreshToken)
+                .orElseThrow(() -> new NoSuchElementException("cannot find refresh token in DB"));
+    }
+}
